gs-odrs-provider: Change parse_success() to take an input stream
This is a prerequisite to changing other parts of the download/parsing
code to parse an input stream. This will improve performance by avoiding
having to allocate a heap buffer for the entire JSON data to be parsed
(which could be megabytes in size).

It introduces no significant changes in itself.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1658