gs-external-appstream-utils: Make system download use gs_plugin_download_file()
This slightly changes the semantics of how the system download works:
previously it would download to a temporary file if the modification
date of the system file was too old. It would then copy the temporary
file to the system file and delete the temporary.

Now, to simplify the code and allow `gs_plugin_download_file()` to be
used, the temporary file is kept around, and it’s used for checking the
ETag/modification date. It’s copied to the system file if modified, but
the temporary file is no longer deleted.

This will allow the system and user code paths to be combined in a
subsequent commit, simplifying things further.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472