gs-plugin: Hold weak references to the plugin in idle callbacks
Various signal emissions from `GsPlugin` are (correctly, I think)
deferred to the global default main context via idle callbacks.

If the plugin was finalised between one of these callbacks being
attached to the main context, and it eventually being invoked, a
use-after-free error would occur.

Avoid that by holding a weak reference to the `GsPlugin` for each idle
callback, and exiting early from the callback if that reference can’t be
converted to a strong one.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1661