fedora-pkgdb-collections: Ensure updates to distros array are immutable
Rather than updating the array in place, atomically replace it with a
new one each time it’s updated. This means that each array is immutable
after construction and can be accessed without a lock (just with a
strong reference).

This is a step towards allowing multiple threads to access the array in
parallel without all holding a lock.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>