packagekit: Auto-prepare update when the file is deleted
The PackageKit prepared update can be deleted in some cases, sometimes
as a side-effect of some actions done by the Software itself, like when
calling refresh after start of the application. Due to the PackageKit
can be used by other applications as well, listen for the changes
on the prepared update file and restore it in case it disappears.

The auto-prepare of the update is triggered only if the file was deleted
and the user has set to download updates. It's not triggered when
the Software is started and there is no prepared update. It's because
the file name is not exposed in the public PackageKit API.