Handle G_IO_ERROR_CANCELLED just like GS_PLUGIN_ERROR_CANCELLED
Eventually I would like to remove `GS_PLUGIN_ERROR_CANCELLED`, as it’s
unhelpful to duplicate a built-in error code from GIO which is
universally used to represent cancellation in *all* the libraries
gnome-software uses.

However, untangling error handling without introducing bugs is not
possible in the amount of time before the release, so this is a
transition step. This is because I can’t be sure there aren’t other bits
of gnome-software which implicitly assume an error is in the
`GS_PLUGIN_ERROR` domain.

On the assumption that some bits of gnome-software will return
`G_IO_ERROR_CANCELLED`, and some will return
`GS_PLUGIN_ERROR_CANCELLED`, handle both equally in all places where
that error code is checked for.

When `GS_PLUGIN_ERROR_CANCELLED` is removed, these checks can all be
collapsed down to a simple check against `G_IO_ERROR_CANCELLED`.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>