snap: Allow calls to be non-interactive
Currently each `SnapdClient` instance is set to always be interactive,
but that’s not how gnome-software works: it has some interactive
(foreground, user requested) and some non-interactive (background, not
user requested) jobs. The snap plugin should forward that information to
the snap daemon so that it doesn’t (for example) show polkit
authentication dialogues for background operations.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>