gs-plugin-job-list-installed-apps: Add flags property
I should have added this when originally writing the class: it provides
a way to say whether the job is interactive.

While `GsPluginJob:interactive` exists as a property, that can’t extend
to the `GsPlugin.list_installed_apps_async()` vfuncs which this job
eventually calls, so the enum needs to be added for them. If it’s being
added there, it might as well be used to set the interactive state on
the plugin job.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>