gs-plugin-loader: Avoid a warning about removing a non-existent source
When running under gdb, the `timeout_id` should be cleared when
returning `G_SOURCE_REMOVE`, just as it is at the end of the function
when *not* running under gdb.

This fixes a warning from `g_source_remove()` about an unknown source ID
when waiting for a long time in gdb.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>