gs-plugin-appstream: Fix some application types not being correctly queried
The 'console' type never existed; for the 'webapp' legacy type there is
already a fixup in place to use the 'web-application' type currently
in use in the specification[1]; and the 'desktop-application' type has been
for a while the default identifier for desktop applications,
super-seeding 'desktop'[2], so its fixup has been added.

[1] https://www.freedesktop.org/software/appstream/docs/sect-Metadata-WebApplication.html
[2] https://www.freedesktop.org/software/appstream/docs/sect-Metadata-Application.html#spec-appdata-filespec

squash! gs-plugin-appstream: fixup legacy 'desktop' to 'desktop-application' type