gs-application: Move plugin setup to the end of the activate vfunc
This means that all the private objects in `GsApplication` will have
been created before the async call starts, which means that other vfunc
calls on the `GsApplication` in the meantime will be able to use them.

In particular, this means that `app->shell` will exist by the time
`activate()` is called (and it’s typically called right after
`startup()`).

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1670