ui: Disable markup in widgets derived from AdwPreferencesRow
This is a follow-up to !1280. If a recent enough version of libadwaita
is available, it disables markup in widgets derived from
`AdwPreferencesRow`, rather than having to explicitly escape all label
contents.

This is not worth bumping our libadwaita dependency for, as it’s a
simple and small set of changes, and libadwaita 1.2.0 is not quite
released yet (so won’t be in distros for a while).

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>