gs-download-utils: Don’t use ETag when writing local file
The ETag we care about for the download is the ETag returned by the
server, not the ETag of the local file. Passing the server’s ETag to the
local file operations may result in the local file not being written to,
with a “The file was externally modified” error.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1677