gs-download-utils: Ignore cancellation errors when there’s a cache hit
In order to avoid overwriting an existing cache file when the server has
said it’s up to date, we cancel closing the output stream.

Avoid propagating that cancellation error up to the caller; to them, the
operation should seem like a success rather than returning
`G_IO_ERROR_CANCELLED`.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1679