gs-plugin-loader: Inherit list properties when replacing it
When running refine on a subset of the apps returned by a plugin action,
the original list is discarded and replaced with the new refined list.
The problem is that the discarded list has set important properties,
like whether it had been randomized, truncated or how many max items
the list had so far.

These properties are usable for example when search for the apps, thus
inherit them before discarding the original list.