fedora-pkgdb-collections: Avoid refreshing the collections early
During early startup of gnome-software, refresh-metadata is called, and
then refine is potentially called on the `install-queue` (if it exists).

refresh-metadata will start refreshing the cache, which will create the
existing cache file to be empty if it doesn’t already exist. If the
refine job is then run before the refresh has completed, it will end up
loading the empty cache file and then erroring out as it fails to parse
it.

Given that the refine job is part of setup, the error will then
propagate and cause setup to fail.

Work around that by avoiding refreshing the cache if the refine job
isn’t relevant to any OS updates (which it’s unlikely to be in any
case). This works around the issue, but is also correct in its own
right, as it spreads work out over time rather than causing it to all be
done immediately at startup.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1670