gs-download-utils: Support If-Modified-Since caching
The download code already has support for ETags, but in case the server
doesn’t support them, or the client doesn’t support the xattrs required
to save an ETag alongside its file, it’s useful to support cache queries
based on the modification time of a file.

Add support for `Last-Modified` and `If-Modified-Since`. This changes
the download APIs, so bumps the API revision.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>