Allow web apps to be in UNAVAILABLE state
After a web app is uninstalled in Software, it is only re-installable if
it is present in the AppStream data. Otherwise the app must have been
installed via Epiphany, and the information needed to re-install it
(e.g. the icon) is not available. If we leave the state as
GS_APP_STATE_UNKNOWN after the removal, the user is greeted by an ugly
message "Unable to find org.gnome.Epiphany.WebApp_...." in place of the
details page. So set the state to GS_APP_STATE_UNAVAILABLE instead, and
update other parts of Software to allow that state to mean "Don't show
any Install/Remove/Open button" when it is set on a web app.