gs-plugin-loader: Add D-Bus connection properties
These will be set either at construction time of the `GsPluginLoader`,
or (by default) when `setup_async()` is called.

In a subsequent commit, they will be passed to the `GsPlugin` objects
the plugin loader loads. The ultimate aim is to allow plugins to easily
use a mock D-Bus connection provided by unit tests.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1694