ci: Run D-Bus session and system buses
They are explicitly needed in all the tests now, as the plugin loader
connects to the bus earlier.

gnome-software already needed both buses (the session bus for
`GApplication`, and the system bus for PackageKit and flatpak), but the
changes in the past few commits mean that those connections are made
earlier in the process, and hence don’t get avoided in some tests.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1694