gs-app-query: Add immutable query object for listing apps
This will be used in upcoming commits, with the `list_apps_async()`
vfunc on `GsPlugin`, to specify which apps should be returned.

It’s a separate object, rather than a series of arguments to
`list_apps_async()`, to make the argument list more manageable and
extensible in future.

It’s immutable so that there are no concerns about thread safety when
using `GsAppQuery` in plugins which use multiple threads. After
construction, it’s read-only.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472