gs-plugin-job-list-apps: Add job class for listing apps
This replaces `GS_PLUGIN_ACTION_SEARCH_FILES` at the moment, and will
eventually replace a number of other actions which query for apps in
different ways.

It’s a separate class so that type safety can be improved, and
behaviour specific to listing apps can be implemented in the class
without complicating `GsPluginLoader` further.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472