gs-app: Make list of addons immutable
Change `gs_app_add_addons()` to use the read-copy-update approach, so
that any lists of addons returned by previous calls to
`gs_app_dup_addons()` are not modified.

While we’re at it, make it operate on a list of addons, rather than a
single addon, which should reduce the number of read-copy-update cycles
needed in `gs-appstream.c`.

Finally, avoid allocating an internal list of addons in a `GsApp` until
the first addon is added. This should avoid carrying an empty allocated
`GsAppList` around in various apps which will never have addons. It
should save a small amount of memory.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1721