flatpak: Always drop internal caches for both installation instances
When a refresh is requested, always drop internal flatpak caches
for both FlatpakInstallation instances, to ensure they both agree
on the local cache content. Otherwise the other FlatpakInstallation
uses stale data and can error out on calls when being used.

An example: the refresh can be called with interactive, but search
with non-interactive, then the search can fail to refine metainfo
data with error "No such ref", because its internal flatpak cache
was not updated.