flatpak: Invalidate caches when app data rescan fails
Otherwise the gnome-software caches may not reflect actual state of
the flatpak data, pretending everything is up-to-date even it's not.

This could happen for example when enabling a flatpak remote
in the Repositories dialog and closing the dialog early, which
cancels the operation, maybe in the middle of downloading appstream
data from the server.