packagekit: Set task background status in task setup
Rather than setting it separately. While `:background` and
`:interactive` have different semantics, we currently always want to set
them using the same conditions.

This means `pk_client_set_background()` is now more consistently called
before a `PkClient` is used; previously it was possible for a stale
background value from the previous use of the `PkClient` to be reused
incorrectly.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>