gs-updates-page: Hide "Operating System Updates Unavailable" by default
It can take some time to get the information about the current operating
system, during which the Updates page shows "Operating System Updates Unavailable"
info bar at the top, which is misleading and confusing. Let's hide it by default
and show it only after the information about the operating system is retrieved.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1743