gs-plugin-loader: Move default category-apps sort function to callers
Callers shouldn’t be relying on a default sort order which is
unspecified. Providing a default like this ties down the API of the
plugin loader.

It’s clearer and more maintainable for callers to set their own sort
functions explicitly, and have no default.

This introduces no functional changes.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472