gs-app-query: Add a helper to count the number of query properties set
Until all the plugins are rewritten to support complex queries which
filter by more than one query property, those plugins need to reject
such complex queries.

Add a helper function to count the number of query properties set, which
will help with this as the number of query properties increases.

This could eventually be removed once all plugins support complex
queries, but that’s unlikely to happen in the near future, as there are
currently no bits of the UI which use such queries.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472