gs-app: Split ‘unknown’ state out of app size values into separate type
Rather than representing an ‘unknown’ size as `0` and an ‘unknowable’
size as `G_MAXUINT64`, this changes all the size properties in `GsApp`
to represent them as `(GS_SIZE_TYPE_UNKNOWN, _)` and
`(GS_SIZE_TYPE_UNKNOWABLE, _)`, and to represent known/valid sizes as
`(GS_SIZE_TYPE_VALID, size)`.

This allows `0` byte sizes to be represented consistently, which is
important for indicating that an app doesn’t need any additional
downloads (if `GsApp:size-download` is `0`, for example).

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>