gs-review-row: Fix visibility of ‘Remove’ button for user’s reviews
The ‘Remove’ button is supposed to be shown for any review the current
user has written. Unfortunately, it was being hidden due to the
`box_voting` box which contains it being hidden when the upvote/downvote
buttons are hidden.

Fix that by adding another box around `box_voting` and moving the Remove
and Report buttons up a level into the new box, so they aren’t hidden
when `box_voting` is.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>