rpm-ostree: Improve wait for transaction end
This improves the wait for an ongoing transaction end by checking
a G_IO_ERROR_BUSY error, returned from the rpm-ostree call, and
waiting for the transaction end again. That covers the race when
something else runs a new transaction just before the rpm-ostree call.

The expected error had been added with:
https://github.com/coreos/rpm-ostree/pull/3691