gs-app-list: Make randomising a list a lot more efficient
The previous implementation iterated over the list generating a
3-character randomisation key for each app, and storing it in a per-app
hash table. Then it did a quicksort (O(n log(n))) of the list. Then it
iterated over the list again and removed the randomisation key from each
app.

This took O(n log(n)) time with a heavy constant factor and a lot of
string and hash table allocations. It used O(3n) calls to the `GRand`.

Shuffling an array only needs to take O(n) time, though. This commit
achieves that by using a Fisher–Yates shuffle, with N calls to `GRand`
and N swaps of elements in the array. No allocations are needed apart
from to set up the `GRand`.

I haven’t measured the performance improvement with this change, but it
should be a positive one even if it’s not significant.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>