gs-appstream: Add support for <supports> element
This is another element in the same family as `<requires>` and
`<recommends>`.

It has a slightly different meaning from `<recommends>`, but for now the
gnome-software code treats it as identical, as it’s new to the appstream
spec and many apps which might want to use it are probably still using
`<recommends>` instead.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1774