gs-appstream: Fix getting apps for categories if first group is empty
Categories in gnome-software each comprise one or more XDG standardised
desktop groups, and are the union of all the apps in those groups.

Previously, the code in `gs_appstream_add_category_apps()` would
incorrectly return if any of the queries for apps in a particular
desktop group returned no results. This meant the remaining groups in
the category would not be checked.

Fix that by continuing on to the next group if a query returns no
results.

Spotted by code inspection.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>