packagekit: Use a new PkClient for each refine operation
Creating a `PkClient` is cheap, and there’s no reason (that I can see)
to reuse a single instance throughout the lifetime of the PackageKit
plugin.

The downside of reusing a single instance is that it means multiple
refine operations potentially block on the associated mutex. That
blocking can be avoided by using per-operation `PkClient` instances.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1762