subprojects: Make gnome-pwa-list a submodule
This way gnome-software can be built from a tarball or from git without
network access at build time, as discussed here:
https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/1616#note_1470108

As discussed on !1387, we keep the `.wrap` file so that Meson knows how
to download the module if someone gets hold of the gnome-software
sources without using `git clone` or downloading a tarball which
includes subprojects. It also means that maintainers don’t have to
remember to pass `--include-subprojects` to `meson dist`.