gs-app: Fix developer comment for gs_app_get_unique_id()
The gtk-doc is failing when parsing gs-app.xml due to opening and closing
elements do not match. Escape the <> properly to avoid the problem.

The asterisk in the text around forward slash is also problematic, and
the backslash escaping doesn't work here, as the backslash makes it into
the generated HTML, thus use a different example unique ID.