fwupd: Fix the user action dialogs
When gnome-software was ported from appstream-glib to libappstream we
missed that the locale arguments are now swapped -- i.e. the former
specified "locale, value" and the latter specified "value, locale".

This fixes installing software on any devices that require some kind of
user-prompting to complete the update, e.g. buttons that need pressing
or cables that need replugging.