build: Drop valgrind option and dependency
It was used to avoid unloading plugins when running under valgrind, so
that stack information could be handled by valgrind in leak reports
correctly.

The previous commit fixes this in a different way, by marking plugin
modules as resident. This allows reference counting them as normal, but
prevents them from actually being unloaded when their reference count
reaches zero.

Given that, we can drop this commit and simplify our dependency logic a
teeny bit.

This is a spiritual revert of commit
0e2e452aa010139eea7ef31a785e563bf9476e56. It can’t be a verbatim revert
as a lot of nearby lines have changed since that commit landed.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>