flatpak: Remove NEW_PERMISSIONS quirk if it doesn’t need to be set
It’s feasible that `gs_flatpak_set_update_permissions()` could be called
twice on the same `GsApp` instance, and end up clearing the permissions
diff the second time round. If so, the quirks should be updated to
match.

It’s possible this fixes #1804, but a bit unlikely since I can’t
actually see a way for `gs_flatpak_set_update_permissions()` to be
called twice on the same `GsApp`.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1804