lib: Drop support for plugin job timeouts
This basically reverts the changes from
https://bugzilla.gnome.org/show_bug.cgi?id=784251. Rather than force a
timeout on plugins which are misbehaving, it’s better to fix the plugins
so they don’t time out.

If plugins are communicating with another service (such as snapd or a
D-Bus service), then timeouts can be implemented inside the plugin for
when that communication fails.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>