gs-plugin-appstream: Ignore refine for apps with no ID
It can happen a	plugin (in my case PackageKit) returns apps for	update,	which do
not have set their ID. Trying to fins such app in an XbSilo resulted in	an error:

   failed to process component/id[text()='(null)']: cannot parse text or number `null`

which prevented the Updates page to show anything. Ignoring such apps does not
influence the System Updates, these bare packages are part of it too.