ui: Don’t fade in spinners which are next to a status label
If these spinners fade in, the labels are visible for a few seconds
without a spinner. During that time, it looks like they’re layed out in
the wrong place.

Fading in the spinner was (I think) an attempt to reduce glitchiness of
briefly showing a spinner for a fraction of a second before the content
loads. However, if a label is being shown as well then there’s already
glitchiness so the fading doesn’t achieve anything.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

See: #1429