gs-app-details-page: Change the title to ‘Update Details’
It looks weird to show the package name and/or version in the dialog
title. Instead, just consistently show ‘Update Details’.

If we later decide that we need to show the package name/version
somewhere, the right place for it is lower down in the dialog, not in
the title bar.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1460