data: Update and expand the default curated apps list
The old list wasn’t long enough to actually populate the overview page
if it was the only source of curated apps. 12 apps are needed for the
section to be shown on the overview page.

While we don’t expect distros to only use this list (or to use this list
at all) – they should be providing their own curated lists – GNOME OS
currently uses this list. We should make it work.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1809