upgrade-banner: Subclass AdwBin
Now that GsUpdatesPage is properly disposed, it revealed that
GsUpgradeBanner was not unparenting its only child, a GtkBox.
This triggers a GTK warning.

This could have been worked around by unparenting this single
box, but libadwaita provides a widget tailored for single
child widgets: AdwBin. It also takes care of disposing for us,
and is generally a better option here.

Make GsUpgradeBanner subclass AdwBin.