gs-plugin-job-refine: Add profiler marks
This one is a bit more involved. The refine operation run each
plugin's refine function asynchronously, but sequencially, so
we have to be careful to set the begin time of each plugin run
individually.

This commit adds the same pair of marks (one for each plugin,
one for the overall job time) but because of the above, it needs
to use two struct fields.