gs-search-page: Ignore finished non-current searches
The search can take various time and even the cancellable is cancelled
due to the search term change or anything like that, the code can be
far away from the place where the cancellable could be checked, which
means even the search was obsolete it could modify the content of the page.

Let the code check whether the finished search is still the current
search and ignore any result from it when it's an obsolete search.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1731