epiphany: Setup without worker thread
Setting up with the worker thread presents a problem for the
Epiphany plugin: if the plugin disables itself, it triggers an
assertion failure on shutdown. The series of events that causes
this goes like this:

 - Software starts up, calls setup_async on all plugins
 - The Epiphany plugin creates a worker thread and runs setup
 - In the worker thread, Epiphany disables itself by returning FALSE
 - During shutdown, GsPluginLoader skips shutting down disabled
   plugins (this skips calling gs_worker_thread_shutdown!)
 - GsPluginLoader clears the plugins array, calling dispose on all
   created plugins
 - GsPluginEpiphany clears the worker thread in dispose
 - GsWorkerThread fails an assertion, since it's being destroyed
   with a running thread

This commit fixes that by making the entire setup process of Epiphany
plugin asynchronous, and only creating the worker thread at the end
of setup.