gs-updates-section: Flip function call order in _unreveal_row()
Even the order technically doesn't matter at the moment, it'll be better
to follow the pattern to first connect to the signal and only then
unreveal the row. That will help to avoid possible use-after-free in case
the call to unreveal the row would emit the "unrevealed" signal immediately
in the future, which the code currently handles as "delete the GsAppRow".