build: Drop libsoup API version check for flatpak library
The `soupapiversion` variable had been added in [1], but it did not
land yet and possibly never will. The flatpak library can be built
with libcurl instead, which should be detectable somehow too. As there
is currently no way to know what libsoup version the flatpak library
had been built with, if any, drop the check in the build script.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1877

[1] https://github.com/flatpak/flatpak/pull/4582