flatpak: Avoid deadlock under gs_flatpak_refine_wildcard()
The gs_flatpak_refine_wildcard() calls with a reader lock on the silo.
That function calls gs_flatpak_refine_app_unlocked(), which calls
ensure_flatpak_silo_with_locker(), which can request a writer lock
due to invalidated silo, which causes the deadlock due to the held
reader lock. Avoid it by propagating the reader lock up to the call
stack and release it before the writer lock is requested.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1748