external-appstream: Read ETag of correct files
gs_download_file_async() is used by the external AppStream code to
download files with extra AppStream data. It supports two modes:

 * User, where external AppStream files are downloaded and
   installed into ~/.cache/gnome-software/external-appstream

 * System, where external AppStream files are downloaded and
   installed into /var/cache/swcatalog/xml

When gs_download_file_async() is called, it first reads the ETag
of the local file, if it exists, and passes that to Soup so that
Soup can compare the ETag against the server file's ETag. This
works well for user installations, since the download folder is
also the installation folder.

For system-wide installations, however, this does not work. That's
because GNOME Software downloads the AppStream file into the user
directory (~/.cache/gnome-software/external-appstream), then moves
the downloaded file into the system folder. In practice, for system
installations, the ETag is always NULL, because the local file never
exists.

Fix that by doing most of what gs_download_file_async() inside of
gs_external_appstream_refresh_async(). Namely, it created necessary
folders; read the ETag of *the correct file*; replaces the local
output file; then finally uses gs_download_stream_async() to fetch
and install the AppStream file.

This allows Soup to compare the correct ETag with the server file,
and skip the download entirely if they match.

This commit also exposes another bug in the external AppStream code
where it leaves behind an empty file when the download is skipped,
which follow-up commits will fix.

Related: https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1902