download-utils: Return new NOT_MODIFIED error if ETags match
The downloading routines of GNOME Software treat matching ETags
as not errors, however, there are circumstances where it'll be
important to react to this result differently from other error
paths.

Add the GS_DOWNLOAD_ERROR quark, and GS_DOWNLOAD_ERROR_NOT_MODIFIED
to represent that case. Next commits will adjust the error paths of
consumers of this API to not do anything, or react appropriately,
if that error is reported.