external-appstream: Delete empty file if not modified on server
On system-wide AppStream installations, if the external AppStream
file has an ETag that matches that of the server, GNOME Software
does not remove the temporary file which is created right before
the download operation started.

Now that we know when such cases happen, through the new error
id GS_DOWNLOAD_ERROR_NOT_MODIFIED, remove the temporary download
file if it's a system-wide installation.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1902