rpm-ostree: Mark apps as downloaded
There was a typo in the changes of the gs_app_set_size_download() API,
the `0` meant the app is downloaded, not that the size is unknown.
With this fixed, the Updates page offers to `Restart & Update`, instead
of `Download`, which does nothing, because the rpm-stree plugin
does not implement corresponding GsPlugin function.