gs-details-page: Allow specify packaging format with origin name
Let the format be optionally extended with an origin name, to be able
to define packaging format in a more fine-grained way. Also let specify
only the origin name, which will match any packaging format. An example is:

   [ 'flatpak:flathub', ':system', 'rpm', 'flatpak' ]

which prefers apps from the Flathub Flatpak remote, if it does not exist,
then chooses an app from origin "system" in any format, if it does not exist,
then chooses an RPM version of the app, and if even that does not exist,
then chooses a Flatpak version from any remote.