gs-install-appstream: Remove old files from legacy path
Since the path the external appstream data files are stored changed,
users with old files updating gnome-software to a version with the changed
path would have left the old files with no chance to have them deleted
other than manually by themselves. This tries to delete the file in
the legacy path first and moves the file to the new location.

Related to https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1909