gs-appstream: Mute "ignoring unknown or empty provided item type" runtime warning
This is a problem in the appstream data, with which the user can do nothing,
thus do not fill his/her journals with things he/she cannot influence nor fix.
These are still shown in the verbose log.

Spotted as:

ignoring unknown or empty provided item type: font

where the node text is `NULL`.