gs-plugin: Verify app belongs to the repo before changing app state
This is for gs_plugin_update_cache_state_for_repository(), which updates
app state based on the repo enable/disable state. Repos for different
plugins can have the same name, thus verify also that the app belongs
to the same plugin and is of the same kind as the changed repo is.

Related to https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1952