contrib: Disable PolicyKit support in the Flatpak manifest
PolicyKit is mostly used by the PackageKit plugin, which is disabled
in the Flatpak build. The Flatpak plugin uses it only to check
whether it can use the system installation, which is not usable
in the Flatpak sandbox, thus disable the PolicyKit in the build.

Also update the build dependencies in the manifest.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1862