gs-app-reviews-dialog: Avoid listbox rows removal when possible
The GtkListBox tries very hard to keep cursor (focused) row in the view,
and whenever any row is removed, it changes the position of the vertical
scrollbar to reflect the change. Better to merge review changes into
existing rows also to keep preserved the cursor row, because removing all
rows and adding there back new rows with the same reviews does lost
the cursor row, not only the vertical scrollbar position.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1776