rpm-ostree: Install pending updates when automatic updates are on
Install pending updates when automatic updates are on, thus the only
thing the updates are waiting on is a machine restart, regardless whether
it's done in the Software or in the system. The 'Restart & Update' button
in the Updates page works more smoothly now and it does just that, it
restarts the machine and applies the updates.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/659