eos-updater: Clear download size once update is fetched
The `GsApp:download-size` property is the size of *what remains* to be
downloaded, not the size of the entire update. That means it needs to be
updated to zero once the update is complete.

This fixes the ‘Download’ button sometimes being shown again after
downloading the update. In particular, this can happen if gnome-software
is started when the eos-updater daemon is part-way through an update
(and hence the gnome-software process has not proceeded through the Poll
phase).

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1985