eos-updater: Rename gs_plugin_app_upgrade_download() implementation
The next few commits will introduce a local call to this function. If
it’s not renamed, and the local call done to a `static` version of the
function, it’s possible that `gs_plugin_app_upgrade_download()` will
resolve to the implementation of that function in another plugin. That
would break things.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>