eos-updater: Support major vs minor OSTree updates in the UI
Use the new `UpdateIsUserVisible` property of eos-updater to determine
whether to show an update in the UI using the OS upgrade banner (if it
contains significant user visible changes) or as an update row like
non-app package updates on a package-based OS (if it doesn’t contain
significant user visible changes).

See https://github.com/endlessm/eos-updater/pull/312 for the changes in
eos-updater.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1543