overview-page: Replace the 3rd-party infobar by a dialog
This was implemented as an infobar rather than a dialog because of the
"Want to go shopping?" dialog, which is now gone. This drops the infobar
and reimplement it as a message dialog, which not only works better but
allows us to get rid of the deprecated GtkInfoBar, using
AdwMessageDialog instead.